requirements.txt
numpy == 2.3.1
torch == 2.7.1
torchvision == 0.22.1
matplotlib == 3.10.3
pillow == 11.2.1